finally fixed weird segfaults
This bug was weird. Rust program was segfaulting. That holds for at least
half the commits before this one. I couldn't locate core of the problem,
because it crashed in random locations during memory deallocation.
First, I thought it was a logger, that was crashing. I changed it to
a different one, problem wasn't fixed and now the new logger was crashing.
It didn't take me long to realise, that the problem was with syscall handling.
But I couln't find the exact spot. I had trouble pinpointing the exact
syscall, that was causing it. I always though, I have found it and then boom,
it appeared again somewhere else.
With the help of Valgrind, I then found out, that it was allways crashing
during deallocation. I thought, it might be a problem with Rust itself and I was
writing a post to the Rust community about the issue. Then I noticed, that Valgrind
always tells me, that is libpthread2.7. What about trying the same thing on a different
system? I did so and it helped me sooo much. On Ubuntu, the program compiled
a bit differently and it was failing with ABORT, not SEGFAULT. Valgrind also
noticed some more issues in a completely different syscall than the one
I thought was the one. And so it struck me.

All the time, it wasn't the syscalls. It was the only other thing that was
unsafe in Rust. Creating CStrings. Because CStrings are owned, they free their
memory. And that was the issue. Because the memory wasn't theirs.

This bug took me about a week to fix. In the end, the fix consists of changing
CString to CStr. Just that. Unsafe code is just unsafe, it's that simple. Don't use it,
when you can get by without it!